Enable build with VS 2017

+++ src/cpp/web-ifc/parsing/IfcTokenStream.cpp
@@ -50,10 +50,10 @@ namespace webifc::parsing
       _fileStream->Clear();
   }
 
-  void IfcTokenStream::SetTokenSource(std::istream &requestData)
-  {
-     SetTokenSource([&](char* dest, size_t sourceOffset, size_t destSize) { requestData.seekg(sourceOffset); requestData.read(dest, destSize); return requestData.gcount();},true);
-  }
+   void IfcTokenStream::SetTokenSource(std::istream &requestData)
+   {
+     SetTokenSource([&](char* dest, size_t sourceOffset, size_t destSize) { requestData.seekg(sourceOffset); requestData.read(dest, destSize); return static_cast<uint32_t>(requestData.gcount());},true);
+   }
 
   std::string_view IfcTokenStream::ReadString()
   {
